fix: make result_t pointer-sized to fix gitclient crash on arm64
result_t was uint32_t (32-bit), which truncated pointers on arm64
when tvGetSelectedRecord returned a record pointer through send_message.
Changed result_t to intptr_t and send_message to return intptr_t.

Also disabled gitclient debug logging by default (GITCLIENT_DEBUG=0).

+++ user/window.c
@@ -87,7 +87,7 @@ ui_runtime_state_t g_ui_runtime = {
 
 // Forward declarations
 extern void post_message(window_t *win, uint32_t msg, uint32_t wparam, void *lparam);
-extern int send_message(window_t *win, uint32_t msg, uint32_t wparam, void *lparam);
+extern intptr_t send_message(window_t *win, uint32_t msg, uint32_t wparam, void *lparam);
 extern int titlebar_height(window_t const *win);
 extern int statusbar_height(window_t const *win);
